- Fixed PSP compiler problems. - Disabled the Settings menu Breakpoint address option on the PSP, due to the compiler crashing compiling it. - Fixed some SDL 1.2 compatibility problems. - Fixed some variable sizes to be more correct.

+++ UniPCemu/hardware/modem.c
@@ -152,7 +152,7 @@ byte readIPnumber(char **x, byte *number)
 {
 	byte size=0;
 	word result=0;
-	for (;(isdigit(*(*x)) && (size<3));) //Scan digits!
+	for (;(isdigit((int)*(*x)) && (size<3));) //Scan digits!
 	{
 		result = (result*10)+(*(*x)-'0'); //Convert to a number!
 		++(*x); //Next digit!
